Rename Bitset parent package from utils to collections (#571)
* rename-bitset-parent-package-to-collections

* Fix Dockerfiles

* Better comments for Bitset encoding

* Gracefully handle nil bitset encoding
