+++ lib/pos_network_test.go
@@ -6,7 +6,7 @@ import (
 	"testing"
 
 	"github.com/deso-protocol/core/bls"
-	"github.com/deso-protocol/core/utils/bitset"
+	"github.com/deso-protocol/core/collections/bitset"
 	"github.com/stretchr/testify/require"
 )
